+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
----------
+=========
 
+
+8.20.0
+------
+* tab for displaying local sessions and their respective status
+* additional task parameters for passiveChoiceWorld
+* work on making the GUI code more modular
 
 8.19.6
 ------
@@ -43,15 +49,21 @@ Changelog
 * add ambient sensor reading back to trial log
 * allow negative stimulus gain (reverse wheel contingency)
 
+-------------------------------
+
 8.18.0
 ------
 * valve calibration routine
+
+-------------------------------
 
 8.17.0
 ------
 * consolidated data transfer routine across all rig computers
 * new video workflow with support for multiple named camera setups
 * various small fixes, work on documentation, unit-tests
+
+-------------------------------
 
 8.16.1
 ------
@@ -63,6 +75,8 @@ Changelog
 * Support for Bpod HiFi Module
 * Support for Zapit Optostim (NM)
 * more robust handling of Bpod's serial messages: iblrig.hardware._define_message
+
+-------------------------------
 
 8.15.6
 ------
@@ -100,6 +114,8 @@ Changelog
 * feature: calibration routine for frame2ttl v1-3 in Tools menu
 * feature: debug-flag for IBLRIG Wizard
 
+-------------------------------
+
 8.14.2
 ------
 * hotfix: wrong return-type in _iterate_protocols - pt 2
@@ -111,6 +127,8 @@ Changelog
 8.14.0
 ------
 * show dialog boxes and plots for appended sessions
+
+-------------------------------
 
 8.13.5
 ------
@@ -146,6 +164,8 @@ Changelog
 * added button for triggering a free reward (only available outside of running task for now)
 * cleaned-up logging
 * various improvements under the hood, clean-up and unit-tests
+
+-------------------------------
 
 8.12.13
 -------